dtEntitySUT branch:
* tick message simulation time was set with float precision instead of double, fixed
* added safety check to manipulator script 
* ground clamper can now be configured to fetch lowest LOD for intersection
* chancing simulation clock string to no longer have date, only time
* Enqueueing cameraCreatedMessage now instead of emitting it directly, this way recipients can be sure that camera entity is fully created
* improved error message in entity tree
* simulation clock time caused an overflow on some computers in scriptcomponent.cpp

+++ branches/dtEntitySUT/include/dtEntity/systemmessages.h
@@ -79,7 +79,7 @@ namespace dtEntity
       void SetDeltaSimTime(float v) { mDeltaSimTime.Set(v); }
       void SetDeltaRealTime(float v) { mDeltaRealTime.Set(v); }
       void SetSimTimeScale(float v) { mSimTimeScale.Set(v); }
-      void SetSimulationTime(float v) { mSimulationTime.Set(v); }
+      void SetSimulationTime(double v) { mSimulationTime.Set(v); }
 
    private:
